antig/pmg-creative-studio/local-services/image-edit-api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.3.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
pytesseract==0.3.13
